+++ crates/holochain-conductor-runtime-ffi/src/runtime.rs
@@ -1,16 +1,11 @@
 use crate::error::RuntimeResultFfi;
+use crate::multi_thread::MultiThreadRun;
 use android_logger::Config;
 use holochain_conductor_runtime::{move_to_locked_mem, Runtime, RuntimeConfig};
 use holochain_conductor_runtime_types_ffi::*;
 use holochain_types::prelude::InstallAppPayload;
 use log::{debug, LevelFilter};
-use std::sync::LazyLock;
-use tokio::runtime::{Builder, Runtime as TokioRuntime};
 use url2::Url2;
-
-/// Global multi threaded tokio runtime
-pub static RT: LazyLock<TokioRuntime> =
-    LazyLock::new(|| Builder::new_multi_thread().enable_all().build().unwrap());
 
 /// Slim wrapper around HolochainRuntime, with types compatible with Uniffi-generated FFI bindings.
 #[derive(uniffi::Object, Clone)]
@@ -70,14 +65,9 @@ impl RuntimeFfi {
     pub async fn install_app(&self, payload: InstallAppPayloadFfi) -> RuntimeResultFfi<AppInfoFfi> {
         debug!("RuntimeFfi::install_app");
         let payload: InstallAppPayload = payload.try_into()?;
+        let res = MultiThreadRun::exec(self.0.install_app(payload)).await?;
 
-        #[cfg(not(test))]
-        let res = RT.block_on(async { self.0.install_app(payload).await });
-
-        #[cfg(test)]
-        let res = self.0.install_app(payload).await;
-
-        Ok(res?.into())
+        Ok(res.into())
     }
 
     /// Uninstall an app
@@ -105,6 +95,30 @@ impl RuntimeFfi {
     ) -> RuntimeResultFfi<AppAuthFfi> {
         debug!("RuntimeFfi::ensure_app_websocket");
         let app_auth = self.0.ensure_app_websocket(installed_app_id).await?;
+
+        Ok(AppAuthFfi {
+            authentication: app_auth.authentication.into(),
+            port: app_auth.port,
+        })
+    }
+
+    /// Full process to setup an app
+    ///
+    /// Check if app is installed, if not install it, then optionally enable it.
+    /// Then ensure there is an app websocket and authentication for it.
+    ///
+    /// If an app is already installed, it will not be enabled. It is only enabled after a successful install.
+    /// The reasoning is that if an app is disabled after that point,
+    /// it is assumed to have been manually disabled in the admin interface, which we don't want to override.
+    pub async fn setup_app(
+        &self,
+        payload: InstallAppPayloadFfi,
+        enable_after_install: bool,
+    ) -> RuntimeResultFfi<AppAuthFfi> {
+        debug!("RuntimeFfi::setup_app");
+        let payload: InstallAppPayload = payload.try_into()?;
+        let app_auth =
+            MultiThreadRun::exec(self.0.setup_app(payload, enable_after_install)).await?;
 
         Ok(AppAuthFfi {
             authentication: app_auth.authentication.into(),
@@ -142,7 +156,7 @@ mod test {
         runtime
             .install_app(InstallAppPayloadFfi {
                 source: HAPP_FIXTURE.to_vec(),
-                installed_app_id: Some(app_id.into()),
+                installed_app_id: app_id.into(),
                 network_seed: Some(Uuid::new_v4().to_string()),
                 roles_settings: Some(HashMap::new()),
             })
@@ -212,7 +226,7 @@ mod test {
         let res = runtime
             .install_app(InstallAppPayloadFfi {
                 source: HAPP_FIXTURE.to_vec(),
-                installed_app_id: Some("my-app-1".into()),
+                installed_app_id: "my-app-1".into(),
                 network_seed: Some(Uuid::new_v4().to_string()),
                 roles_settings: Some(HashMap::new()),
             })
@@ -448,5 +462,37 @@ mod test {
         let res = runtime.enable_app("non-existant-app-1".into()).await;
         assert!(res.is_err());
         assert!(matches!(res, Err(RuntimeErrorFfi::Runtime { .. })))
+    }
+
+    #[tokio::test(flavor = "multi_thread")]
+    async fn test_setup_app() {
+        let tmp_dir = TempDir::new().unwrap();
+        let tmp_dir_path = tmp_dir.path().as_os_str().to_str().unwrap().to_string();
+        let runtime = RuntimeFfi::start(
+            vec![0, 0, 0, 0],
+            RuntimeConfigFfi {
+                data_root_path: tmp_dir_path,
+                bootstrap_url: "https://bootstrap.holo.host".into(),
+                signal_url: "wss://sbd.holo.host".into(),
+            },
+        )
+        .await
+        .unwrap();
+
+        let res = runtime
+            .setup_app(
+                InstallAppPayloadFfi {
+                    source: HAPP_FIXTURE.to_vec(),
+                    installed_app_id: "my-app-1".into(),
+                    network_seed: Some(Uuid::new_v4().to_string()),
+                    roles_settings: Some(HashMap::new()),
+                },
+                true,
+            )
+            .await;
+        assert!(res.is_ok());
+
+        let apps = runtime.list_apps().await.unwrap();
+        assert_eq!(apps.len(), 1);
     }
 }