+++ crates/tauri-plugin-holochain-service-consumer/src/mobile.rs
@@ -39,8 +39,10 @@ impl<R: Runtime> HolochainServiceConsumer<R> {
     /// 5. Ensure there is an app websocket and return its authentication config
     ///
     /// This can be safely called whether or not the app is installed.
-    pub fn setup_app(&self, config: SetupAppConfig) -> crate::Result<AppAuthFfi> {
-        Ok(self.0.run_mobile_plugin::<AppAuthFfi>("setupApp", config)?)
+    pub fn connect_setup_app(&self, config: SetupAppConfig) -> crate::Result<AppAuthFfi> {
+        Ok(self
+            .0
+            .run_mobile_plugin::<AppAuthFfi>("connectSetupApp", config)?)
     }
 
     /// Build a window that opens the main UI for your Tauri app.
@@ -77,15 +79,15 @@ impl<R: Runtime> HolochainServiceConsumer<R> {
     }
 
     /// Setup a holochain app, then create the main window builder which will inject the holochain env JS on initialization.
-    /// 
+    ///
     /// Note you must call `.build()` on the returned WebviewWindowBuilder to actually build the window.
     pub fn setup_app_main_window(
         &self,
         config: SetupAppConfig,
     ) -> tauri::Result<WebviewWindowBuilder<R, AppHandle<R>>> {
-        let app_auth = self.setup_app(config.clone()).ok();
+        let app_auth = self.connect_setup_app(config.clone()).ok();
         let window_builder = self.main_window_builder(config.app_id, app_auth.clone())?;
-        
+
         Ok(window_builder)
     }
 }